made some major ish chnages
added error handling for 404, 500, and, 405.  added a flash page so that dynamic messages based off of app.py can be sent to the user

added the profile path to the sql db

fixed logout function so that users can loggout of their account

adjusted the nave bar to reflect if a user is logged in or not.
    users not logged in cannot see logout or profile,

users logged in cannot se login or create account.